+++ test/libsolidity/smtCheckerTests/types/mapping_struct_assignment.sol
@@ -11,8 +11,4 @@ contract C
 	}
 }
 // ----
-// Warning 8364: (159-160): Assertion checker does not yet implement type type(struct C.S storage pointer)
-// Warning 4639: (159-163): Assertion checker does not yet implement this expression.
 // Warning 6838: (140-144): BMC: Condition is always false.
-// Warning 8364: (159-160): Assertion checker does not yet implement type type(struct C.S storage pointer)
-// Warning 4639: (159-163): Assertion checker does not yet implement this expression.