+++ test/libsolidity/smtCheckerTests/types/struct/struct_nested_temporary.sol
@@ -1,0 +1,18 @@
+pragma experimental SMTChecker;
+
+contract C {
+
+	struct S {
+		uint x;
+	}
+
+	struct T {
+		S s;
+		uint y;
+	}
+
+	function test() pure public {
+		assert(T(S(42), 1).s.x == 42);
+	}
+}
+// ----